+++ mex/common/models/__init__.py
@@ -98,6 +98,10 @@ from mex.common.models.activity import (
     PreviewActivity,
     SubtractiveActivity,
 )
+from mex.common.models.base.container import (
+    ItemsContainer,
+    PaginatedItemsContainer,
+)
 from mex.common.models.base.extracted_data import ExtractedData
 from mex.common.models.base.filter import BaseFilter, FilterField, FilterRule
 from mex.common.models.base.mapping import BaseMapping, MappingField, MappingRule
@@ -353,6 +357,7 @@ __all__ = (
     "ExtractedVariableGroup",
     "FilterField",
     "FilterRule",
+    "ItemsContainer",
     "MappingField",
     "MappingRule",
     "MergedAccessPlatform",
@@ -377,6 +382,7 @@ __all__ = (
     "OrganizationalUnitMapping",
     "OrganizationalUnitRuleSetRequest",
     "OrganizationalUnitRuleSetResponse",
+    "PaginatedItemsContainer",
     "PersonFilter",
     "PersonMapping",
     "PersonRuleSetRequest",